Split study,py into _study_summary.py and study.py to avoid the circular reference.
