+++ optuna/study.py
@@ -11,6 +11,7 @@ from joblib import Parallel
 
 from optuna._experimental import experimental
 from optuna._study_direction import StudyDirection
+from optuna._study_summary import StudySummary
 
 try:
     import pandas as pd  # NOQA
@@ -48,101 +49,6 @@ if type_checking.TYPE_CHECKING:
 
 
 _logger = logging.get_logger(__name__)
-
-
-class StudySummary(object):
-    """Basic attributes and aggregated results of a :class:`~optuna.study.Study`.
-
-    See also :func:`optuna.study.get_all_study_summaries`.
-
-    Attributes:
-        study_name:
-            Name of the :class:`~optuna.study.Study`.
-        direction:
-            :class:`~optuna.study.StudyDirection` of the :class:`~optuna.study.Study`.
-        best_trial:
-            :class:`FrozenTrial` with best objective value in the :class:`~optuna.study.Study`.
-        user_attrs:
-            Dictionary that contains the attributes of the :class:`~optuna.study.Study` set with
-            :func:`optuna.study.Study.set_user_attr`.
-        system_attrs:
-            Dictionary that contains the attributes of the :class:`~optuna.study.Study` internally
-            set by Optuna.
-        n_trials:
-            The number of trials ran in the :class:`~optuna.study.Study`.
-        datetime_start:
-            Datetime where the :class:`~optuna.study.Study` started.
-    """
-
-    def __init__(
-        self,
-        study_name,  # type: str
-        direction,  # type: StudyDirection
-        best_trial,  # type: Optional[structs.FrozenTrial]
-        user_attrs,  # type: Dict[str, Any]
-        system_attrs,  # type: Dict[str, Any]
-        n_trials,  # type: int
-        datetime_start,  # type: Optional[datetime]
-        study_id,  # type: int
-    ):
-        # type: (...) -> None
-
-        self.study_name = study_name
-        self.direction = direction
-        self.best_trial = best_trial
-        self.user_attrs = user_attrs
-        self.system_attrs = system_attrs
-        self.n_trials = n_trials
-        self.datetime_start = datetime_start
-        self._study_id = study_id
-
-    def __eq__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, StudySummary):
-            return NotImplemented
-
-        return other.__dict__ == self.__dict__
-
-    def __lt__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, StudySummary):
-            return NotImplemented
-
-        return self._study_id < other._study_id
-
-    def __le__(self, other):
-        # type: (Any) -> bool
-
-        if not isinstance(other, StudySummary):
-            return NotImplemented
-
-        return self._study_id <= other._study_id
-
-    @property
-    def study_id(self):
-        # type: () -> int
-        """Return the study ID.
-
-        .. deprecated:: 0.20.0
-            The direct use of this attribute is deprecated and it is recommended that you use
-            :attr:`~optuna.study.StudySummary.study_name` instead.
-
-        Returns:
-            The study ID.
-        """
-
-        message = (
-            "The use of `StudySummary.study_id` is deprecated. "
-            "Please use `StudySummary.study_name` instead."
-        )
-        warnings.warn(message, DeprecationWarning)
-
-        logger = logging.get_logger(__name__)
-        logger.warning(message)
-
-        return self._study_id
 
 
 class BaseStudy(object):
@@ -984,7 +890,7 @@ def delete_study(
 
 
 def get_all_study_summaries(storage):
-    # type: (Union[str, storages.BaseStorage]) -> List[structs.StudySummary]
+    # type: (Union[str, storages.BaseStorage]) -> List[StudySummary]
     """Get all history of studies stored in a specified storage.
 
     Args:
@@ -993,7 +899,7 @@ def get_all_study_summaries(storage):
             :func:`~optuna.study.create_study` for further details.
 
     Returns:
-        List of study history summarized as :class:`~optuna.structs.StudySummary` objects.
+        List of study history summarized as :class:`~optuna.study.StudySummary` objects.
 
     """
